+++ workbench/routes/web.php
@@ -6,22 +6,10 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Route;
 use Workbench\App\Models\User;
 
-Route::get('/config', function () {
-    dump(config('legacy-bridge'));
-});
-
-Route::get('/clear-session', function () {
-    Auth::logout();
-
-    Cookie::queue(Cookie::forget('legacy-test-session'));
-
-    DB::table('legacy_sessions')->where('id', 'legacy-test-session')->delete();
-
-    return redirect('/protected');
-});
+Route::get('/login', fn () => 'login')->middleware('guest')->name('login');
+Route::get('/', fn () => Auth::guard()->check() ? 'Logged in!' : 'Not logged in...')->middleware('auth');
 
 Route::get('/create-legacy-session', function () {
-
     DB::table('legacy_sessions')->insertOrIgnore([
         'id'            => 'legacy-test-session',
         'payload'       => base64_encode(serialize(['user_id' => User::query()->value('id')])),
@@ -40,8 +28,19 @@ Route::get('/create-legacy-session', function () {
         config('session.same_site') ?? null
     );
 
-    return redirect('/protected');
+    return redirect('/');
 });
 
-Route::get('/login', fn () => 'login')->name('login');
-Route::get('/protected', fn () => Auth::guard()->check() ? 'Logged in!' : 'Not logged in...')->middleware('auth');
+Route::get('/config', function () {
+    dump(config('legacy-bridge'));
+});
+
+Route::get('/clear-session', function () {
+    Auth::logout();
+
+    Cookie::queue(Cookie::forget('legacy-test-session'));
+
+    DB::table('legacy_sessions')->where('id', 'legacy-test-session')->delete();
+
+    return redirect('/');
+});