+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.1.2 - 2026-07-09
+
+### Changed
+
+- `php artisan legacy-bridge:verify` - more useful outputs
+- `LegacyDatabaseSessionHandler` - small refactor
+
 ## 0.1.1 - 2026-07-08
 
 ### Added